fix (internal/schedulers/time/scheduler): refactor functions, update dbclient with pkg/client/dbserver
