+++ internal/schedulers/time/scheduler/init.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"time"
 
-	"github.com/trigg3rX/triggerx-backend/internal/schedulers/time/client"
 	"github.com/trigg3rX/triggerx-backend/internal/schedulers/time/config"
 	"github.com/trigg3rX/triggerx-backend/internal/schedulers/time/metrics"
+	"github.com/trigg3rX/triggerx-backend/pkg/client/dbserver"
 	"github.com/trigg3rX/triggerx-backend/pkg/client/aggregator"
 	"github.com/trigg3rX/triggerx-backend/pkg/logging"
 	"github.com/trigg3rX/triggerx-backend/pkg/types"
@@ -17,7 +17,7 @@ type TimeBasedScheduler struct {
 	cancel                  context.CancelFunc
 	logger                  logging.Logger
 	activeTasks             map[int64]*types.ScheduleTimeTaskData
-	dbClient                *client.DBServerClient
+	dbClient                *dbserver.DBServerClient
 	aggClient               *aggregator.AggregatorClient
 	metrics                 *metrics.Collector
 	schedulerSigningAddress string
@@ -30,7 +30,7 @@ type TimeBasedScheduler struct {
 }
 
 // NewTimeBasedScheduler creates a new instance of TimeBasedScheduler
-func NewTimeBasedScheduler(managerID string, logger logging.Logger, dbClient *client.DBServerClient, aggClient *aggregator.AggregatorClient) (*TimeBasedScheduler, error) {
+func NewTimeBasedScheduler(managerID string, logger logging.Logger, dbClient *dbserver.DBServerClient, aggClient *aggregator.AggregatorClient) (*TimeBasedScheduler, error) {
 	ctx, cancel := context.WithCancel(context.Background())
 
 	scheduler := &TimeBasedScheduler{
@@ -66,4 +66,46 @@ func NewTimeBasedScheduler(managerID string, logger logging.Logger, dbClient *cl
 	)
 
 	return scheduler, nil
+}
+
+// Start begins the scheduler's main polling and execution loop
+func (s *TimeBasedScheduler) Start(ctx context.Context) {
+	s.logger.Info("Starting time-based scheduler", "scheduler_signing_address", s.schedulerSigningAddress)
+
+	ticker := time.NewTicker(s.pollingInterval)
+	defer ticker.Stop()
+
+	for {
+		select {
+		case <-ctx.Done():
+			s.logger.Info("Scheduler context cancelled, stopping")
+			return
+		case <-s.ctx.Done():
+			s.logger.Info("Scheduler stopped")
+			return
+		case <-ticker.C:
+			s.pollAndScheduleTasks()
+		}
+	}
+}
+
+// Stop gracefully stops the scheduler
+func (s *TimeBasedScheduler) Stop() {
+	startTime := time.Now()
+	s.logger.Info("Stopping time-based scheduler")
+
+	// Capture statistics before shutdown
+	activeTasksCount := len(s.activeTasks)
+
+	s.cancel()
+
+	duration := time.Since(startTime)
+
+	s.logger.Info("Time-based scheduler stopped",
+		"duration", duration,
+		"active_tasks_stopped", activeTasksCount,
+		"performer_lock_ttl", s.performerLockTTL,
+		"task_cache_ttl", s.taskCacheTTL,
+		"duplicate_task_window", s.duplicateTaskWindow,
+	)
 }
